+++ src/bbt-io.ads
@@ -50,6 +50,7 @@ private package BBT.IO is
                             Spawn,
                             Step_Actions,
                             Step_Lexer,
+                            Step_Parser,
                             Tests_Builder,
                             Writers);
    subtype Topics is Extended_Topics range